+++ CHANGELOG.md
@@ -24,7 +24,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - [feature] Assign a post ID to multiple reviews in bulk
 
 ### Fixed
-- Fix system info PHP error when detecting server IP address
+- Fix CSS for themes that do not implement the `.screen-reader-text` class
+- Fix System Info PHP error when detecting server IP address
 
 ## [2.8.4] = 2017-11-16