+++ CHANGELOG.md
@@ -18,6 +18,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Security
 
+## [2.9.1] = 2017-12-08
+
+### Fixed
+- Fix "read more" links with ajax pagination
+
 ## [2.9.0] = 2017-12-08
 
 ### Added